Split :juxt.card.alpha/content into content and children
This is a key refactoring that untangles the two different uses of the use of
the :juxt.card.alpha/content keyword.

A new keyword, :juxt.card.alpha/children, is introduced to create a document's
structure, leaving :juxt.card.alpha/content to describe the segments of a
paragraph.

This removes some difficulty and confusion in understanding the code.